refactor(server): split schema definitions from migration code
Keep the versioned Hot Updater schema as the primary maintenance surface while preserving the existing adapter API and migration outputs.

Constraint: FumaDB is removed, but db generate/migrate must remain schema-first and provider-compatible.

Rejected: Keep all schema DSL, version snapshots, SQL diffing, and Mongo operations in hotUpdaterSchema.ts | makes future schema changes hard to review.

Confidence: high

Scope-risk: moderate

Directive: Add future schema versions in packages/server/src/db/schema/definitions.ts first; keep provider implementation code in sibling schema modules.

Tested: pnpm -w lint:fix; pnpm exec tsc -p packages/server/tsconfig.json --noEmit; pnpm exec vitest run packages/server/src/db/index.spec.ts packages/hot-updater/src/commands/generate.spec.ts packages/hot-updater/src/commands/migrate.spec.ts --reporter=dot; pnpm exec vitest run plugins/plugin-core/src/bundleUnitOfWork.spec.ts plugins/plugin-core/src/createDatabasePlugin.spec.ts packages/server/src/db/pluginCore.spec.ts packages/server/src/handler-standalone.integration.spec.ts --reporter=dot; pnpm --filter @hot-updater/server build

Not-tested: Standalone agent matrix after this follow-up commit is still pending.
