+++ packages/server/src/adapters/drizzle.ts
@@ -32,7 +32,7 @@ import {
 import {
   getHotUpdaterSchemaVersion,
   hotUpdaterSchema,
-} from "../db/hotUpdaterSchema";
+} from "../db/schema/registry";
 import { generateDrizzleSchema } from "../db/schemaGenerators";
 import type {
   DatabasePluginFactory,
